+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fixed an issue which prevented the promotion process of additional Active Directory Domain
   Controllers from being restartable (#1608).
 - Fixed a syntax issue in 'CreateDscSqlDatabase.ps1'.
+- Fixed an issue in Stop-LabVm (#1619) with invalid error objects.
 
 ## 5.50.0 (2023-11-10)